hw2/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pyjwt==2.10.0
bcrypt==4.2.1
python-json-logger==3.2.0
python-multipart==0.0.20
pydantic==2.10.3
pydantic-settings==2.6.1
